logthrsource: publish LogThreadedSourceWorker
